requirements.txt
pytest
pytest_bdd
selenium
pymysql
pony
jsonpickle
allure-pytest